fix: avoid math.MaxInt64 uint overflow on 32-bit platforms
isIntegerType's uint case compared val against math.MaxInt64, which
overflows uint on 32-bit platforms (e.g. linux/arm), breaking the
cross-compile. Convert to uint64 for the bounds check.

+++ filters/standard_filters.go
@@ -36,8 +36,10 @@ func isIntegerType(v any) bool {
 	case int, int8, int16, int32, int64, uint8, uint16, uint32:
 		return true
 	case uint:
-		// Check if uint value fits in int64 range
-		return val <= math.MaxInt64
+		// Check if uint value fits in int64 range.
+		// Convert to uint64 so the comparison constant doesn't overflow
+		// uint on 32-bit platforms (e.g. linux/arm).
+		return uint64(val) <= math.MaxInt64
 	case uint64:
 		// Check if uint64 value fits in int64 range
 		return val <= math.MaxInt64
